requirements.txt
discord.py
discord.py[voice]
captcha
tzdata
python-dotenv
tortoise-orm[asyncpg]
aerich

aivoice-python
robyn
requests
jinja2